Python/requirements.txt
# RFSN GenAI Orchestrator v8.2 - Python Dependencies
# Production-grade with pinned versions

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# LLM Inference
llama-cpp-python==0.2.56

# JWT
PyJWT==2.8.0

# HTTP requests (for model downloads)
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Static Files
aiofiles==23.2.1

# Learning Layer
numpy>=1.24

# TTS (MLX)
mlx>=0.0.10
mlx-audio-plus
huggingface_hub[hf_xet]
scipy
soundfile
librosa
einops
mlx_lm

# Semantic Memory
sentence-transformers==2.2.2
scikit-learn>=1.3.0
